+++ changelog.md
@@ -6,6 +6,7 @@
 - Move config/value_property to platforms.decorators
 - Move measure_execution_time to decorators
 - Move definition exporter to device
+- Move HM_INTERFACE_EVENT_SCHEMA to central
 - Update requirements
 
 # Version 2023.9.1 (2023-09-06)